+++ main.py
@@ -1,14 +1,10 @@
-import torch
-from visualization import plot_tours, plot_loss_curve
-from data import get_tsp_dataloader
 import argparse
-from utils import (
+from helpers import (
     find_latest_checkpoint,
-    save_model,
     get_experiment_dir,
-    get_device,
 )
-from builder import build_model, build_optimizer_and_scheduler
+from utils import train, evaluate, rl_fine_tune_with_dpo
+from builder import build_model
 import logging
 import os
 import time
@@ -33,226 +29,13 @@ def setup_logging(experiment_dir, log_level):
     return logging.getLogger(__name__)
 
 
-def calculate_avg_solution_tour_length(dataloader, device):
-    """
-    Calculate the average solution tour length from the validation dataloader.
-
-    Args:
-        dataloader (DataLoader): DataLoader for the validation dataset.
-        device (torch.device): The device to run the calculation on.
-
-    Returns:
-        float: The average solution tour length.
-    """
-    total_solution_tour_length = 0.0
-    num_tours = 0
-    for batch in dataloader:
-        batch = batch.to(device)
-        for problem in batch:
-            total_solution_tour_length += problem.solution_tour_length.item()
-            num_tours += 1
-
-    return total_solution_tour_length / num_tours
-
-
-def train(solver, experiment_dir, num_epochs, max_lr, use_scheduler, logger):
-    """
-    Train the solver model on the TSP dataset.
-
-    Args:
-        solver (TransformerSolver): The solver model to be trained.
-        experiment_dir (str): Directory to store experiment results and logs.
-        num_epochs (int): Number of training epochs.
-        max_lr (float): Maximum learning rate.
-        use_scheduler (bool): Whether to use the learning rate scheduler.
-        logger (logging.Logger): Logger for logging training information.
-    """
-    device = get_device()
-
-    solver.to(device)
-
-    # Get the training and validation dataloaders
-    # Create load_path based on the parameters
-    num_samples = 51200
-    min_points = 10
-    max_points = 15
-    train_seed = 42
-    val_seed = 21
-
-    train_load_path = f"data/train/problems_{num_samples}_{min_points}_{max_points}_{train_seed}.pkl"
-    val_load_path = f"data/val/problems_{num_samples}_{min_points}_{max_points}_{val_seed}.pkl"
-
-    # Update the dataloaders with the new load_paths
-    train_dataloader = get_tsp_dataloader(
-        batch_size=2048,
-        num_samples=num_samples,
-        min_points=min_points,
-        max_points=max_points,
-        seed=train_seed,
-        load_path=train_load_path,
-    )
-    val_dataloader = get_tsp_dataloader(
-        batch_size=2048,
-        num_samples=int(num_samples / 10),
-        min_points=min_points,
-        max_points=max_points,
-        seed=val_seed,
-        load_path=val_load_path,
-    )
-
-    # Build the optimizer and scheduler
-    optimizer, scheduler = build_optimizer_and_scheduler(
-        solver, max_lr, use_scheduler, len(train_dataloader), num_epochs
-    )
-
-    # Training loop
-    losses = []
-    val_losses = []
-    val_tour_lengths = []
-    for epoch in range(num_epochs):
-        epoch_loss = 0.0
-
-        start_time = time.time()
-        with tqdm(total=len(train_dataloader), desc=f"Epoch {epoch + 1}/{num_epochs}", leave=False) as pbar:
-            for batch in train_dataloader:
-                batch = batch.to(device)
-                optimizer.zero_grad()
-                loss = solver(batch)
-                loss.backward()
-                optimizer.step()
-                if use_scheduler:
-                    scheduler.step()
-                epoch_loss += loss.item()
-                pbar.set_postfix(loss=loss.item())
-                pbar.update(1)  # Update the progress bar
-
-        epoch_time = time.time() - start_time
-
-        # Step the learning rate scheduler
-        current_lr = scheduler.get_last_lr()[0] if scheduler else max_lr
-
-        avg_epoch_loss = epoch_loss / len(train_dataloader)
-        losses.append(avg_epoch_loss)
-
-        if (epoch + 1) % 10 == 0:
-            avg_val_loss, avg_tour_length = validate(solver, val_dataloader, device, experiment_dir, epoch)
-            # Save the model checkpoint
-            save_model(solver, epoch, experiment_dir)
-            val_losses.append(avg_val_loss)
-            val_tour_lengths.append(avg_tour_length)
-            logger.info("-" * 50)  # This will log a line of 50 dashes as a separator
-            logger.info(f"Epoch [{epoch+1}/{num_epochs}]")
-            logger.info(f"Epoch Time: {epoch_time:.2f} seconds")
-            logger.info(f"Average Train Loss: {avg_epoch_loss:.4f}")
-            logger.info(f"Validation Loss: {avg_val_loss:.4f}")
-            logger.info(f"Validation Tour Length: {avg_tour_length:.4f}")
-            logger.info(f"Current learning rate: {current_lr:.6f}")
-
-    # Calculate the average solution tour length from the validation dataloader
-    avg_solution_tour_length = calculate_avg_solution_tour_length(val_dataloader, device)
-    logger.info(f"Average Solution Tour Length: {avg_solution_tour_length:.4f}")
-
-    # Add the average solution tour length to the list for plotting
-    val_solution_tour_lengths = [avg_solution_tour_length] * len(val_tour_lengths)
-
-    # Plot the loss curve
-    plot_loss_curve(losses, val_losses, val_tour_lengths, val_solution_tour_lengths, experiment_dir)
-
-    # Plot the tours
-    solver.eval()
-    tours = solver.solve(batch)
-    plot_tours(batch, tours, experiment_dir)
-
-
-def validate(solver, val_dataloader, device, experiment_dir, epoch):
-    """
-    Validate the solver on the validation dataset.
-
-    Args:
-        solver (TransformerSolver): The solver model to be validated.
-        val_dataloader (DataLoader): DataLoader for the validation dataset.
-        device (torch.device): The device to run the validation on.
-
-    Returns:
-        tuple: A tuple containing:
-            - avg_val_loss (float): The average validation loss.
-            - avg_tour_length (float): The average tour length on the validation set.
-    """
-    val_loss = 0.0
-    total_tour_length = 0.0
-    num_tours = 0
-    solver.eval()  # Set the model to evaluation mode
-    with torch.no_grad():  # Disable gradient calculation for validation
-        for batch in tqdm(val_dataloader, desc="Validating", leave=False):
-            batch = batch.to(device)
-            loss = solver(batch)
-            val_loss += loss.item()
-            tours = solver.solve(batch)
-
-            # Calculate tour lengths using batch operations
-            tour_points = torch.gather(batch.points, 1, tours.unsqueeze(-1).expand(-1, -1, 2))
-            tour_lengths = torch.sum(
-                torch.norm(tour_points[:, 1:] - tour_points[:, :-1], dim=2) * batch.padding_mask.float(), dim=1
-            )
-
-            # Apply padding mask to tour lengths
-            total_tour_length += tour_lengths.sum().item()
-
-            # Compute the number of valid tours
-            num_tours += batch.padding_mask.size(0)
-
-    avg_val_loss = val_loss / len(val_dataloader)
-    avg_tour_length = total_tour_length / num_tours
-    os.makedirs(os.path.join(experiment_dir,'visualization'), exist_ok=True)
-    plot_tours(batch, tours, os.path.join(experiment_dir,'visualization', f'epoch_{epoch}_tour.png'))
-    solver.train()
-    return avg_val_loss, avg_tour_length
-
-
-def evaluate(solver, experiment_dir, logger):
-    """
-    Evaluate the solver on the validation dataset.
-
-    Args:
-        solver (TransformerSolver): The solver model to be evaluated.
-        experiment_dir (str): Directory to store experiment results and logs.
-    """
-
-    # Determines the device (CPU/GPU) to run the evaluation on.
-    device = get_device()
-
-    solver.to(device)
-
-    # Load the validation dataset
-    num_samples = 51200
-    min_points = 10
-    max_points = 15
-    val_seed = 21
-
-    val_load_path = f"data/val/problems_{num_samples}_{min_points}_{max_points}_{val_seed}.pkl"
-    val_dataloader = get_tsp_dataloader(
-        batch_size=16,
-        num_samples=num_samples,
-        min_points=min_points,
-        max_points=max_points,
-        seed=val_seed,
-        load_path=val_load_path,
-    )
-    # Validate the solver
-    avg_val_loss, avg_tour_length = validate(solver, val_dataloader, device, experiment_dir, epoch=-1)
-    avg_solution_tour_length = calculate_avg_solution_tour_length(val_dataloader, device)
-    logger.info(
-        f"Validation Loss: {avg_val_loss:.4f}, Validation Tour Length: {avg_tour_length:.4f}, Validation Solution Tour Length: {avg_solution_tour_length:.4f}"
-    )
-
-
 if __name__ == "__main__":
 
     parser = argparse.ArgumentParser(description="Train or evaluate the TSP solver")
     parser.add_argument(
         "--mode",
         type=str,
-        choices=["train", "evaluate"],
+        choices=["train", "evaluate", "rl_fine_tune"],
         default="train",
         help="Select whether to train or evaluate the model (default: train)",
     )
@@ -273,17 +56,25 @@ if __name__ == "__main__":
 
     experiment_dir = get_experiment_dir(args.experiment_name)
 
-    # Sets up logging for the evaluation process.
-    logger = setup_logging(experiment_dir, args.log_level)
-
     if args.load_checkpoint:
         latest_checkpoint = find_latest_checkpoint(experiment_dir)
     else:
         latest_checkpoint = None
 
+    if args.mode == "rl_fine_tune":
+        experiment_dir = experiment_dir + "_rl_fine_tune"
+
+    # Sets up logging for the evaluation process.
+    logger = setup_logging(experiment_dir, args.log_level)
+
+    # Builds the model.
     solver = build_model(logger, latest_checkpoint)
 
     if args.mode == "train":
         train(solver, experiment_dir, args.max_epochs, args.max_lr, args.use_scheduler, logger)
     elif args.mode == "evaluate":
         evaluate(solver, experiment_dir, logger)
+    elif args.mode == "rl_fine_tune":
+        rl_fine_tune_with_dpo(solver, experiment_dir, args.max_epochs, args.max_lr, args.use_scheduler, logger)
+    else:
+        raise ValueError(f"Invalid mode: {args.mode}")
